+++ shoutem.push-notifications/app/app.js
@@ -1,11 +1,4 @@
-import _ from 'lodash';
-
-import {
-  Alert,
-  AppState,
-} from 'react-native';
-
-import { I18n } from 'shoutem.i18n';
+import { AppState } from 'react-native';
 
 import { ext } from './const';
 
@@ -13,6 +6,7 @@ import {
   selectPushNotificationGroups,
   userNotified,
   requestPushPermission,
+  showPushNotification,
 } from './redux';
 
 const DEFAULT_PUSH_NOTIFICATION_GROUP = 'broadcast';
@@ -38,30 +32,10 @@ const appDidMount = (app) => {
     return state[ext()].lastNotification;
   }
 
-  function onNotificationAction(notificationContent) {
-    const action = _.get(notificationContent, 'action');
-
-    if (!!action) {
-      store.dispatch(action);
-    }
-  }
-
   function displayPushNotificationMessage(notification) {
     const notificationContent = notification.content;
-    store.dispatch(userNotified());
-
-    if(notificationContent.openedFromTray) {
-      return onNotificationAction(notificationContent);
-    }
-
-    Alert.alert(
-      I18n.t(ext('messageReceivedAlert')),
-      notificationContent.body,
-      [
-        { text: I18n.t(ext('messageReceivedAlertView')), onPress: onNotificationAction.bind(null, notificationContent) },
-        { text: I18n.t(ext('messageReceivedAlertDismiss')), onPress: () => {} },
-      ]
-    );
+    store.dispatch(userNotified());    
+    store.dispatch(showPushNotification(notificationContent));
   }
 
   store.subscribe(() => {
